Move the helpers and retry logic to separate files
